backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
git+https://github.com/openai/CLIP.git
qdrant-client==1.9.1
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.2
pillow==10.3.0
imageio==2.34.1
aiofiles==23.2.1
pydantic==2.7.1
python-dotenv==1.0.1
